+++ test/test_req_rep.py
@@ -1,9 +1,10 @@
 import unittest
-import logging
 from multiprocessing import Process
 
 from nanoservice import Service
 from nanoservice import Client
+
+from nanoservice import encoder
 
 
 class BaseTestCase(unittest.TestCase):
@@ -16,8 +17,9 @@ class BaseTestCase(unittest.TestCase):
 
     def start_service(self, addr):
         s = Service(addr)
-        s.register('divide', lambda x,y: x/y)
+        s.register('divide', lambda x, y: x/y)
         s.start()
+
 
 class TestTCPProtocol(BaseTestCase):
 
@@ -30,14 +32,42 @@ class TestTCPProtocol(BaseTestCase):
         return res, err
 
     def test_req_rep_w_success(self):
-        res, err = self.make_req(6,2)
+        res, err = self.make_req(6, 2)
         self.assertEqual(3, res)
         self.assertTrue(err is None)
 
     def test_req_rep_w_error(self):
-        res, err = self.make_req(6,0)
+        res, err = self.make_req(6, 0)
         self.assertTrue(res is None)
         self.assertTrue(err is not None)
+
+
+class TestErrors(BaseTestCase):
+
+    def make_req(self, *args):
+        proc = Process(target=self.start_service, args=(self.addr,))
+        proc.start()
+        self.client = Client(self.addr)
+
+        # Change encoder to force service to fail on encoding
+        # since the service uses a MsgPack encoder
+        self.client.encoder = encoder.JSONEncoder()
+
+        # Build and send payload to service to trigger decoding
+        payload = self.client.build_payload('divide', args)
+        self.client.send(payload)
+
+        # Change back to msg pack encoder to read the service response
+        self.client.encoder = encoder.MsgPackEncoder()
+
+        out = self.client.receive()
+        proc.terminate()
+        return out
+
+    def test_decode_error(self):
+        out = self.make_req(6, 2)
+        # response shoud be empty since there was a decode error
+        self.assertEqual('', out)
 
 
 if __name__ == '__main__':